// types/Property.ts
export interface PropertyVideo {
  title: string;
  url: string;
}

export interface Property {
  id: number;

  slug: string;

  title: string;

  description: string;

  propertyType:
    | "Apartment"
    | "Villa"
    | "Plot"
    | "Commercial";

  transactionType:
    | "Sale"
    | "Rent";

  locality: string;

  city: string;

  videos: PropertyVideo[];

  price: number;

  priceLabel: string;

  securityDeposit: number;

  commercialType: string;

  area: number;

  areaUnit:
    | "Sq.ft"
    | "Sq.yd"
    | "Acre"
    | "Hectare";

  bedrooms: number;

  bathrooms: number;

  floorNumber: number;

  washroomType?:
    | "Private"
    | "Shared";

  /**
   * Legacy single image.
   * Kept for backward compatibility.
   */
  featuredImage: string;

  /**
   * Multiple uploaded images.
   */
  images: string[];

  amenities: string[];

  featured: boolean;

  /**
   * Indicates whether the property
   * is generated by the Demo Generator.
   */
  is_demo: boolean;
}